feat(config): enforce localhost-only binding with preference guardrails
Implement Story 1.2 - Localhost binding defaults and preferences guardrails
to prevent accidental network exposure in the no-auth MVP.

Changes:
- Add loopback-only host validation (localhost, 127.0.0.1, ::1)
- Write back corrected values to preferences to keep UI in sync
- Add bind failure handling with actionable user notification
- Add comprehensive unit tests for host/port validation

Acceptance Criteria:
- AC1: Default binding to localhost:61169 with proper notification ✓
- AC2: Empty/whitespace host sanitized to localhost ✓
- AC3: Non-loopback hosts rejected with warning ✓
- AC4: Valid port changes trigger graceful restart ✓
- AC5: Bind failures show clear error notification ✓

+++ src/test/java/io/github/fabb/wigai/config/PreferencesBackedConfigManagerAtddTest.java
@@ -62,11 +62,11 @@ class PreferencesBackedConfigManagerAtddTest {
         when(mockPreferences.getNumberSetting(
             eq("MCP Port"),
             eq("Network Settings"),
-            eq(1024),
-            eq(65535),
-            eq(1),
+            eq(1024.0),
+            eq(65535.0),
+            eq(1.0),
             eq(""),
-            eq(AppConstants.DEFAULT_MCP_PORT)
+            eq((double) AppConstants.DEFAULT_MCP_PORT)
         )).thenReturn(mockPortSetting);
 
         when(mockHostSetting.get()).thenReturn("localhost");
